🚚 rename: 카테고리 유틸 이름 변경
