+++ apis/records/records.js
@@ -2,6 +2,7 @@ import { apiClient } from "../apiClient";
 
 export const fetchRecords = async () => {
   const response = await apiClient.get("/records/");
+  console.log("fetchRecords api >>>", response);
   return response.data;
 };
